+++ js/modules/constants.js
@@ -1,0 +1,28 @@
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра.',
+  'В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают.',
+  'Как можно было поймать такой неудачный момент?!'
+];
+
+const NAMES = ['Иван', 'Александр', 'Михаил', 'Дмитрий', 'Артем', 'Николай', 'Сергей', 'Екатерина',
+  'Анастасия', 'Анна', 'Мария', 'Ольга', 'Татьяна', 'Елена', 'Светлана'];
+
+const DESCRIPTIONS = [
+  'Прекрасный закат на океане',
+  'Величественные горы в снежных вершинах',
+  'Романтическая прогулка по цветочным полям',
+  'Таинственный лес с лучами солнца',
+  'Удивительные оттенки осени в парке',
+  'Милая кошка, сидящая на окне',
+  'Облака в форме сказочного замка',
+  'Уютная деревенская избушка',
+  'Темный лес с пеленой тумана',
+  'Экзотические пальмы на пляже'
+];
+
+export {MESSAGES, NAMES, DESCRIPTIONS};
