+++ js/modules/random.js
@@ -1,0 +1,19 @@
+const getRandomInteger = (min, max) =>
+  Math.floor(Math.random() * (max - min + 1) + min);
+
+const getRandomArrayElement = (elements) =>
+  elements[getRandomInteger(0, elements.length - 1)];
+
+const getRandomId = (min, max) => {
+  const previousValues = [];
+  return function () {
+    let currentValue = getRandomInteger(min, max);
+    while (previousValues.includes(currentValue)) {
+      currentValue = getRandomInteger(min, max);
+    }
+    previousValues.push(currentValue);
+    return currentValue;
+  };
+};
+
+export { getRandomInteger, getRandomArrayElement, getRandomId };
